make it better

+++ food/models.py
@@ -38,7 +38,7 @@ class Dishes(models.Model):
     name = models.CharField(max_length=120)
     description = models.TextField()
     img = models.ImageField(upload_to="static/imgs", null=True)
-    price = models.IntegerField()
+    price = models.DecimalField()
     restaurant = models.ForeignKey(Restaurant, null=True, on_delete=models.CASCADE, related_name="dishes")
 
     def __str__(self):
